refactor(worker): modularize lora training processor
